FEAT: 08-05-25 17:06

+++ app/src/webgl/shaders/vertex-main.ts
@@ -1,6 +1,6 @@
 export default `#version 300 es
-precision highp float;
-precision highp int;
+precision mediump  int;
+precision mediump  float;
 
 layout(location=0) in vec3 a_position;
 layout(location=1) in vec2 a_uv;
